pyproject.toml
[tool.poetry]
name = "replug"
version = "0.1.0"
description = ""
authors = ["Pedro Rodriguez <par@fb.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
transformers = {git = "https://github.com/huggingface/transformers.git", rev = "117a0f6afa3e19d40cb7d19f645f475244219b71"}
pandas = "^2.0.0"
tqdm = "^4.65.0"
pydantic = "^1.10.7"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
openai = "^0.27.4"
datasets = "^2.11.0"
submitit = "^1.4.5"
typer = "^0.7.0"
scipy = "^1.10.1"
requests = "^2.28.2"
faiss-gpu = "^1.7.2"
torch = "^2.0.0"
nltk = "^3.8.1"
sentencepiece = "^0.1.97"
omegaconf = "^2.3.0"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "^8.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
